% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerBeta.R
\name{centerBeta}
\alias{centerBeta}
\title{Center correlations}
\usage{
centerBeta(r, v, w, nu)
}
\arguments{
\item{r}{The correlation to center}

\item{v}{The first parameter of the Beta distribution}

\item{w}{The second parameter of the Beta distribution}

\item{nu}{The parameter of the target Beta distribution}
}
\value{
The centered correlation
}
\description{
The null distribution model is r ~ 2 Beta(v,w) - 1
The centering (target) distribution is r_centered ~ 2 Beta(nu, nu) - 1
}
