% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateShape.R
\name{estimateBetas}
\alias{estimateBetas}
\title{Estimate Beta parameters}
\usage{
estimateBetas(
  corSet,
  left = 0,
  right = 0,
  plot = FALSE,
  fine = NULL,
  trim = 0.01,
  ...
)
}
\arguments{
\item{corSet}{Table of correlation coefficients.}

\item{left}{Left parameter for the Beta distribution.}

\item{right}{Right parameter for the Beta distribution.}

\item{plot}{If TRUE, a plot of the distribution is shown.}

\item{fine}{Number of bins for the histogram.}

\item{trim}{Proportion of the data to be trimmed.}
}
\value{
A vector with the two Beta parameters.
}
\description{
This function estimates the Beta parameters for a distribution given the correlations.
}
