% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidDistance.R
\name{euclidDistance}
\alias{euclidDistance}
\title{Calculate Euclidean distance}
\usage{
euclidDistance(mat, v = 1, w = 1, type = "unsigned")
}
\arguments{
\item{mat}{A matrix of correlation coefficients}

\item{v}{The first parameter of the beta distribution}

\item{w}{The second parameter of the beta distribution}

\item{type}{The type of distance to calculate}
}
\value{
A matrix of distances
}
\description{
Calculates distance network from correlation coefficient matrix using a Euclidean function.
}
