% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerDistance.R
\name{powerDistance}
\alias{powerDistance}
\title{Calculate power distance}
\usage{
powerDistance(r, k, unsigned = TRUE)
}
\arguments{
\item{r}{Correlation coefficient(s), -1 <= r <= 1.}

\item{k}{The power parameter.}

\item{unsigned}{If TRUE, this function computes distances based on unsigned associations (|r|), otherwise it computes distances based on signed correlations (r).}
}
\value{
A matrix of distances
}
\description{
Calculates distances for WCNA from correlation coefficient(s) using a power function.
}
\details{
Correlation and distance are transformed as follows: for type = "unsigned"= 1 - abs(r)^k; for type = "signed" = ((1 - r)/2)^k
}
