% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoidDistance.R
\name{sigmoidDistance}
\alias{sigmoidDistance}
\title{Calculate sigmoid distance}
\usage{
sigmoidDistance(mat, alpha, tau0, type = "unsigned")
}
\arguments{
\item{mat}{A matrix of correlation coefficients}

\item{alpha}{The alpha parameter of the sigmoid function}

\item{tau0}{The tau0 parameter of the sigmoid function}

\item{type}{The type of distance to calculate}
}
\value{
A matrix of distances
}
\description{
Calculates distance network from correlation coefficient matrix using a sigmoid function.
}
