% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardcor-package.R
\docType{package}
\name{standardcor-package}
\alias{standardcor}
\alias{standardcor-package}
\title{standardcor: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pricelab.github.io/standardcor/}
}

}
\keyword{internal}
