% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerDistance.R
\name{powerDistance}
\alias{powerDistance}
\title{Calculate power distance}
\usage{
powerDistance(mat, k, type = "unsigned")
}
\arguments{
\item{mat}{A matrix of correlation coefficients}

\item{k}{The power to raise the distance to}

\item{type}{The type of distance to calculate. Either "unsigned" or "signed"}
}
\value{
A matrix of distances
}
\description{
Calculates distance network from correlation coefficient matrix using a power function.
}
\details{
Correlation and distance are transformed as follows: for type = "unsigned"= 1 - abs(r)^k; for type = "signed" = 1 - r)/2)^k
}
