% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseSpearmanCor2.R
\name{SparseSpearmanCor2}
\alias{SparseSpearmanCor2}
\title{Sparse Spearman Correlation}
\usage{
SparseSpearmanCor2(X, Y = NULL, cov = FALSE)
}
\arguments{
\item{X}{A ranked sparse matrix}

\item{Y}{A second ranked sparse matrix. If Y = NULL, correlations between the columns of X and itself will be taken. If Y is specified, the association between the columns of X and the columns of Y will be calculated.}

\item{cov}{when TRUE the covariance matrix is returned.}
}
\value{
A sparse matrix with ranks of non-zero entries
}
\description{
Replace non-zero entries in a sparse entries with non-zero ranks
}
\details{
This method creates a rank matrix for a sparse matrix X using the following approach:
1. Use non-zero enries in a column to calculate the ranks
2. Add (z-1)/2 to the ranks (only non-zero entries are changed). z is the number of zeros
in the column
Since all the entries are shifted by the same constant (the zeros
are already shifted), the covariance matrix of this shifted matrix is
the same as the rank matrix of the entire matrix (where the zeros would
all also have a rank = (z+1)/2) where z is the number of zeros

This rank matrix can then be used to calculate pearson correlation
(pearson correlation )
 Copied without code changes, and only non-textual change to the comments, from the public blog repository
 https://github.com/saketkc/blog/blob/main/2022-03-10/SparseSpearmanCorrelation2.ipynb
owned by Saket Choudhary of Mumbai, India
23 Mar 2024
}
