% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeFromModel.R
\name{standardizeFromModel}
\alias{standardizeFromModel}
\title{Build a standardized correlation matrix using null models for each pair of component datasets}
\usage{
standardizeFromModel(modelL, analyteL, v.std = 32)
}
\arguments{
\item{modelL}{A list-of-lists structure providing two inputs for each pair of 'Omics datasets $A$ and $B$: a matrix ${\rm modelL}\[\[A\]\]\[\[B\]\]\[\['cor'\]\]$ of raw Spearman correlation values, and shape parameters ${\rm modelL}\[\[A\]\]\[\[B\]\]\[\['shape'\]\] == {\rm c}\(v,w\)$ specifying a null model $(1+r_{raw})/2 ~ {\rm Beta}(v,w)$ for the raw correlations.}

\item{v.std}{}

\item{analtyeL}{For each 'Omics dataset $A$, ${\rm analyteL}\[\[A\]\]$ lists the analytes provided by dataset $A$. These analyte identifiers are required to be unique across all datasets.}
}
\value{
A symmetric matrix containing standardized Spearman correlation coefficients for every pair of analytes across all datasets.
}
\description{
Uses the provided list of null models (${\rm modelL}$) to standardize each type of
correlations to the target common null model specified by the v.std parameter,
$(r + 1)/2 ~ {\rm Beta}({\rm v.std}, {\rm v.std})$, where $r$ is the
standardized Spearman correlation coefficient between any two analytes
in the modeled list of 'Omics datasets. These analtyes are listed by
dataset in the components of the list parameter analtyeL, and must
be unique and contain all analytes provided in each raw correlation matrix
in ${\rm modelL}$. The resulting standardized correlation matrix is built blockwise and is full symmetric.
}
